feat:Solve the size adaptation problem, add GIF function

+++ load_font_face.go
@@ -30,7 +30,7 @@ func loadFontFace() (font.Face, error) {
 		}
 
 		fontFace = truetype.NewFace(f, &truetype.Options{
-			Size: fontPoints,
+			Size: float64(fontPoints),
 		})
 	})
